Update RoommateCard.tsx: edit card color
