+++ src/App.tsx
@@ -7,7 +7,6 @@ import { AnimalDictionary } from './pages/animaldict/AnimalDictionary'
 import { Verify } from './pages/verify/Verify'
 import { ChatList,  } from './pages/chat/ChatList'
 import { RoommateRecommend } from './pages/roommatelist/RoommateRecommend'
-import { RoommateOption } from './pages/roommatelist/RoommateOption'
 import { Group } from './pages/group/Group'
 import { Notice } from './pages/notice/Notice'
 import BottomNavbar from './components/BottomNavBar/BottomNavBar'
