+++ src/forpractice/ForPractice.tsx
@@ -1,10 +1,5 @@
-
-import { Loading } from "../components/Loading"
-import { Intro2 } from "../pages/roommatelist/Intro2"
-import { RoommateOption } from "../pages/roommatelist/RoommateOption"
-import { Card73 } from "../pages/roommatelist/components/Card73"
-export function ForPractice() {
+export const ForPractice = () => {
     return (
-    <div> <Loading  address = 'd'firstLine='모글리님의' secondLine='생활 유형을 분석 중이에요'/></div>
+        <div>ssf</div>
     )
 }
